refactor(sema): inline natspec lowering
